display letters

+++ bitoperations.cpp
@@ -35,6 +35,6 @@ uint8_t changeBit(uint8_t &byte, uint8_t bit, uint8_t value){
     return byte ^= (-value ^ byte) & (1UL << bit);
 }
 
-uint8_t getBit(uint8_t byte, uint8_t bit){
+uint8_t getBit(uint64_t byte, uint8_t bit){
     return (byte >> bit) & 1U;
 }
