+++ bitoperations.h
@@ -21,6 +21,6 @@ uint8_t toggleBit(uint8_t &byte, uint8_t bit);
 
 uint8_t changeBit(uint8_t &byte, uint8_t bit, uint8_t value);
 
-uint8_t getBit(uint8_t byte, uint8_t bit);
+uint8_t getBit(uint64_t byte, uint8_t bit);
 
 #endif // BITOPERATIONS_H
